+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `CodebaseDescriberAgent` to describe the code snippets and enrich the codebase index with contextualized information. This will improve the quality of the search results and the ability of `CodebaseChatAgent` to answer questions.
+
 ### Changed
 
 - Improved `RunSandboxCodeTool` prompt to avoid using it only to print stuff.